Part 1- Changes

+++ Space-Invaders/source/UI/MainMenu/MainMenuUIController.cpp
@@ -60,7 +60,10 @@ namespace UI
         {
             sf::RenderWindow* game_window = ServiceLocator::getInstance()->getGraphicService()->getGameWindow();
 
-            background_image->initialize(Config::background_texture_path, game_window->getSize().x, game_window->getSize().y, sf::Vector2f(0, 0));
+            background_image->initialize(Config::background_texture_path,
+                static_cast<float>(game_window->getSize().x),
+                static_cast<float>(game_window->getSize().y),
+                sf::Vector2f(0, 0));
             background_image->setImageAlpha(background_alpha);
         }
